Create custom hook to fetch games
